+++ src/java_cp1_javier_guerra/controllers/MediaController.java
@@ -1,4 +1,4 @@
-package java_cp1_javier_guerra;
+package java_cp1_javier_guerra.controllers;
 
 import java_cp1_javier_guerra.entities.*;
 import java_cp1_javier_guerra.services.*;
@@ -9,7 +9,7 @@ import java.util.Set;
 import static java_cp1_javier_guerra.entities.MediaFormat.getMediaFormat;
 import static java_cp1_javier_guerra.utils.ConsoleInput.*;
 
-public abstract class MenuOptions {
+public abstract class MediaController {
 
     private static final MediaService mediaService = new MediaServiceImpl();